Removed unnecessary load/store operations when MSAA enabled. This saves lot of memory bandwidth and reduces power consumption
Added new command buffer after Ambient Occlusion and Before ForwardOpaque to control load/store operations
This is mainly used to save memory bandwidth which is caused due to MSAA. It resolves the MultiSampled(4x) Color attachment before storing to memory and don't store the Pre-Resolved
Color attachment which saves a lot of memory bandwidth and helps in low power consumption. Also it doesn't load/store depth attachment which is not necessary as we are using Camera depth texture to get the depth data.