MV3-8642 Resolved black background on the Wheat Nomad scene
The issue was introduced with the recent bandwidth optimisation. To reduce bandwidth we were not loading previous stage color and depth data and the RT was cleared with default color. so it was appearing black.
Now I changed it to  Camera background color so there won't be any black background.
This should not impact the recent optimisation.